+++ src/main/java/com/jamjam/bookjeok/domains/orderdetail/dto/OrderDetailDTO.java
@@ -1,4 +1,4 @@
-package com.jamjam.bookjeok.domains.order.dto.orderdetail;
+package com.jamjam.bookjeok.domains.orderdetail.dto;
 
 import lombok.Builder;
