+++ src/main/java/com/jamjam/bookjeok/domains/orderdetail/service/OrderDetailService.java
@@ -1,7 +1,7 @@
-package com.jamjam.bookjeok.domains.order.service.orderdetail;
+package com.jamjam.bookjeok.domains.orderdetail.service;
 
-import com.jamjam.bookjeok.domains.order.dto.orderdetail.OrderDetailDTO;
-import com.jamjam.bookjeok.domains.order.dto.pendingorder.request.PendingOrderBookItemsRequest;
+import com.jamjam.bookjeok.domains.orderdetail.dto.OrderDetailDTO;
+import com.jamjam.bookjeok.domains.pendingorder.dto.request.PendingOrderBookItemsRequest;
 import com.jamjam.bookjeok.domains.order.entity.Order;
 
 import java.util.List;
