--- src/main/java/com/jamjam/bookjeok/domains/payment/dto/OrderItemsDTO.java
@@ -1,9 +1,0 @@
-package com.jamjam.bookjeok.domains.payment.dto;
-
-import lombok.Builder;
-
-@Builder
-public record OrderItemsDTO(
-        Long bookId, String bookName, int stockQuantity, int totalPrice
-) {
-}
