+++ src/main/resources/mappers/order/orderdetail/OrderDetailMapper.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.jamjam.bookjeok.domains.order.repository.orderdetail.mapper.OrderDetailMapper">
-    <select id="findOrderDetailByOrderId" resultType="com.jamjam.bookjeok.domains.order.dto.orderdetail.OrderDetailDTO">
+<mapper namespace="com.jamjam.bookjeok.domains.orderdetail.repository.mapper.OrderDetailMapper">
+    <select id="findOrderDetailByOrderId" resultType="com.jamjam.bookjeok.domains.orderdetail.dto.OrderDetailDTO">
         SELECT
             o.order_id, b.book_name, b.isbn, od.total_price,
             od.quantity, b.image_url, o.ordered_at
